admin: fix warnings and status code in /load API
