+++ caddyconfig/configadapters.go
@@ -19,6 +19,7 @@ import (
 	"fmt"
 
 	"github.com/caddyserver/caddy/v2"
+	"github.com/caddyserver/caddy/v2/caddyconfig/warning"
 )
 
 // Adapter is a type which can adapt a configuration to Caddy JSON.
@@ -28,20 +29,7 @@ type Adapter interface {
 }
 
 // Warning represents a warning or notice related to conversion.
-type Warning struct {
-	File      string `json:"file,omitempty"`
-	Line      int    `json:"line,omitempty"`
-	Directive string `json:"directive,omitempty"`
-	Message   string `json:"message,omitempty"`
-}
-
-func (w Warning) String() string {
-	var directive string
-	if w.Directive != "" {
-		directive = fmt.Sprintf(" (%s)", w.Directive)
-	}
-	return fmt.Sprintf("%s:%d%s: %s", w.File, w.Line, directive, w.Message)
-}
+type Warning = warning.Warning
 
 // JSON encodes val as JSON, returning it as a json.RawMessage. Any
 // marshaling errors (which are highly unlikely with correct code)
